+++ models/buytoken.js
@@ -3,8 +3,8 @@ module.exports = (sequelize, DataTypes) => {
   var BuyToken = sequelize.define('BuyToken', {
   	user_id:  DataTypes.INTEGER,
     walletMethod: DataTypes.STRING,
-    amount: DataTypes.INTEGER,
-    tokens: DataTypes.INTEGER,
+    amount: DataTypes.DECIMAL(30,30),
+    tokens: DataTypes.DECIMAL(30,30),
     buyHash: DataTypes.STRING,
     buyStatus: DataTypes.ENUM('Pending','Success','Failed'),
     userAddress:DataTypes.STRING
