+++ models/tokentransfer.js
@@ -4,9 +4,9 @@ module.exports = (sequelize, DataTypes) => {
   	user_id:  DataTypes.INTEGER,
     fromAddress: DataTypes.STRING,
     toAddress: DataTypes.STRING,
-    totalTokens: DataTypes.INTEGER,
-    fromToken: DataTypes.INTEGER,
-    toToken: DataTypes.INTEGER,
+    totalTokens: DataTypes.DECIMAL(30,30),
+    fromToken: DataTypes.DECIMAL(30,30),
+    toToken: DataTypes.DECIMAL(30,30),
     transHash:DataTypes.STRING,
     transStatus:DataTypes.ENUM('Pending','Success','Failed'),
   }, {});
